Split mirage resolvers into distinct options
